Split REPL from overall CLI
